requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
arabic-reshaper==3.0.0
arrow==1.3.0
asgiref==3.8.1
asn1crypto==1.5.1
asttokens==2.4.1
attrs==24.2.0
backoff==2.2.1
binaryornot==0.4.4
botocore==1.35.68
Brotli==1.1.0
canvas==0.6.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
configparser==7.1.0
cookiecutter==2.6.0
cron-converter==1.2.1
cryptography==43.0.3
cssselect2==0.7.0
decorator==5.1.1
Django==5.1.2
django-stubs==5.1.1
django-stubs-ext==5.1.1
env-tools==2.3.0
et_xmlfile==2.0.0
executing==2.1.0
fonttools==4.55.0
fpdf==1.7.2
frozenlist==1.5.0
gql==3.5.0
graphql-core==3.2.5
grpcio==1.68.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httpx==0.27.2
idna==3.10
ipython==8.29.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jedi==0.19.2
Jinja2==3.1.4
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keyring==25.5.0
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
more-itertools==10.5.0
multidict==6.1.0
mypy==1.13.0
mypy-extensions==1.0.0
numpy==2.1.3
openpyxl==3.1.5
oscrypto==1.3.0
pandas==2.2.3
parso==0.8.4
pdfkit==1.0.0
pillow==10.4.0
plum-dispatch==1.7.4
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==4.25.5
psycopg==3.2.3
psycopg-binary==3.2.3
pure_eval==0.2.3
pycparser==2.22
pydantic==2.10.1
pydantic_core==2.27.1
pydyf==0.11.0
Pygments==2.18.0
pyHanko==0.25.3
pyhanko-certvalidator==0.26.5
PyJWT==2.10.0
pypdf==5.1.0
pyphen==0.15.0
python-bidi==0.6.3
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2024.2
pywin32-ctypes==0.2.3
PyYAML==6.0.2
qrcode==8.0
RapidFuzz==3.10.1
redis==5.2.0
referencing==0.35.1
reportlab==4.2.5
requests==2.32.3
requests-toolbelt==1.0.0
RestrictedPython==7.4
rich==13.9.4
rpds-py==0.21.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
Spire.Xls==14.7.3
sqlparse==0.5.1
stack-data==0.6.3
statsd==4.0.1
svglib==1.5.1
text-unidecode==1.3
tini==3.0.1
tinycss2==1.4.0
tinyhtml5==2.0.0
traitlets==5.14.3
typer==0.13.1
types-python-dateutil==2.9.0.20241003
types-PyYAML==6.0.12.20240917
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
uritools==4.0.3
urllib3==2.2.3
wcwidth==0.2.13
weasyprint==63.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==11.0.3
xhtml2pdf==0.2.16
yarl==1.18.0
zopfli==0.2.3.post1
